fix: share profile not closing the modal.
