+++ github/handlers/ExecuteBlockActionHandler.ts
@@ -289,71 +289,6 @@ export class ExecuteBlockActionHandler {
                     await this.persistence.updateByAssociation(new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, "ProfileShareParam"), storeData);
                     break;
                 }
-                case ModalsEnum.SHARE_PROFILE_EXEC : {
-                    let {user, room} = context.getInteractionData();
-                    const block = this.modify.getCreator().getBlockBuilder();
-                    let accessToken = await getAccessTokenForUser(this.read, user ,this.app.oauth2Config) as IAuthData;
-                    const userProfile = await getBasicUserInfo(this.http, accessToken.token);
-
-                    const idRecord = new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, "ProfileShareParam")
-
-                    const profileData = await this.read.getPersistenceReader().readByAssociation(idRecord);
-
-                    let profileShareParams: string[] = [];
-
-                    if (profileData.length == 0){
-                        profileShareParams = ['username', 'avatar', 'email', 'bio', 'followers', 'following' , 'contributionGraph'];
-                    }
-                    else {
-                        const dat = profileData[0] as {profileParams : string[]};
-                        profileShareParams = dat.profileParams;
-                    }
-
-                    if (profileShareParams.includes('avatar')){
-                        block.addImageBlock({
-                            imageUrl : userProfile.avatar,
-                            altText : "User Info"
-                        })
-                    }
-
-                    profileShareParams.map((value) => {
-                        if (value != 'contributionGraph' && value != 'avatar'){
-                            block.addSectionBlock({
-                                text : block.newPlainTextObject(value),
-                            })
-                            block.addContextBlock({
-                                elements : [
-                                    block.newPlainTextObject(userProfile[value], true),
-                                ]
-                            });
-                            block.addDividerBlock();
-                        }
-                    })
-
-                    if (profileShareParams.includes('contributionGraph')){
-                        block.addImageBlock({imageUrl : `https://activity-graph.herokuapp.com/graph?username=${userProfile.username}&bg_color=ffffff&color=708090&line=24292e&point=24292e`, altText: "Github Contribution Graph"})
-                    }
-
-
-                    if(user?.id){
-                        if(room?.id){
-                            await sendMessage(this.modify, room!, user, `${userProfile.name}'s Github Profile`, block)
-                        }else{
-                            let roomId = (
-                                await getInteractionRoomData(
-                                    this.read.getPersistenceReader(),
-                                    user.id
-                                )
-                            ).roomId;
-                            room = await this.read.getRoomReader().getById(roomId) as IRoom;
-                            await sendMessage(this.modify, room, user, `${userProfile.name}'s Github Profile`, block)
-                        }
-                    }
-
-                    this.persistence.removeByAssociation(idRecord);
-
-                    break;
-                }
                 case ModalsEnum.VIEW_FILE_ACTION: {
                     const codeModal = await fileCodeModal({
                         data,
